# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/assets": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Lists assets",
                "operationId": "listAssets",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetPagingResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Adds a new asset",
                "operationId": "storeAsset",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssetStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assets/{id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get Asset information",
                "description": "Returns Asset data",
                "operationId": "showAsset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update Asset information",
                "description": "Returns updated Asset data",
                "operationId": "updateAsset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssetStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete Asset",
                "description": "Deletes a single Asset",
                "operationId": "deleteAsset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Asset": {
                "title": "Asset",
                "description": "Asset model",
                "type": "object",
                "xml": {
                    "name": "Asset"
                },
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseModel"
                    },
                    {
                        "properties": {
                            "fqdn": {
                                "title": "fqdn",
                                "description": "FQDN of the asset",
                                "type": "string",
                                "format": "string",
                                "example": "example.com"
                            },
                            "unique_id": {
                                "title": "unique_id",
                                "description": "The unique id of the Asset",
                                "type": "string",
                                "format": "string",
                                "example": "1"
                            },
                            "operating_system": {
                                "title": "operating_system",
                                "description": "The operating system of the Asset",
                                "type": "string",
                                "format": "string",
                                "example": "Windows"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "BaseModel": {
                "title": "BaseModel",
                "description": "BaseModel",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BaseModel"
                }
            },
...
            "AssetStoreRequest": {
                "title": "AssetStoreRequest",
                "properties": {
                    "fqdn": {
                        "title": "fqdn",
                        "description": "FQDN of the new Asset",
                        "type": "string",
                        "example": "example.com"
                    },
                    "unique_id": {
                        "title": "unique_id",
                        "description": "The unique id of the Asset",
                        "type": "string",
                        "example": "1"
                    },
                    "operating_system": {
                        "title": "operating_system",
                        "description": "The operating system of the Asset",
                        "type": "string",
                        "example": "Windows"
                    }
                },
                "type": "object"
            },
...
            "AssetPagingResource": {
                "title": "AssetPagingResource",
                "description": "AssetPagingResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Asset"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/Meta"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AssetPagingResource"
                }
            },
            "AssetResource": {
                "title": "AssetResource",
                "description": "AssetResource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Asset"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AssetResource"
                }
            },
...
            "Meta": {
                "title": "Meta",
                "properties": {
                    "current_page": {
                        "title": "Current Page",
                        "description": "Current Page of the data",
                        "type": "integer"
                    },
                    "from": {
                        "title": "From",
                        "description": "From",
                        "type": "integer"
                    },
                    "last_page": {
                        "title": "Last Page",
                        "description": "Last Page of the data",
                        "type": "integer"
                    },
                    "per_page": {
                        "title": "Per Page",
                        "description": "Per Page of the data",
                        "type": "integer"
                    },
                    "to": {
                        "title": "To",
                        "description": "To",
                        "type": "integer"
                    },
                    "total": {
                        "title": "Total",
                        "description": "Total",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format Bearer 4|oeXad4kChJT43wli90LOd1VbFhtuGuEdvxvEHMAtcb025185",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}